update polkadex to spec 273

+++ app/src/parser_impl.h
@@ -174,7 +174,7 @@ parser_error_t _readEra(parser_context_t *c, pd_ExtrinsicEra_t *v);
 
 parser_error_t _readTx(parser_context_t *c, parser_tx_t *v);
 
-uint8_t _getAddressType();
+uint16_t _getAddressType();
 
 parser_error_t _toStringCompactInt(const compactInt_t *c, uint8_t decimalPlaces,
                                    char postfix[],
